feat: 채팅방 들어가기 + 거래상태변경
