+++ client/src/components/views/PostListPage/PostDetail.jsx
@@ -10,8 +10,8 @@ import socket from "../../../socket";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
 import "swiper/css/pagination";
-//import { Pagination } from "swiper";
-import { Pagination } from "swiper/modules";
+import { Pagination } from "swiper";
+//import { Pagination } from "swiper/modules";
 import { createGlobalStyle } from "styled-components";
 
 const GlobalStyle = createGlobalStyle`
